widen return address in dynarec

+++ src/Core/r4300/recomp.cpp
@@ -26,7 +26,7 @@ precomp_block *dst_block;     // the current block that we are recompiling
 uint32_t src;                 // the current recompiled instruction
 int32_t fast_memory;
 
-uint32_t *return_address; // that's where the dynarec will restart when
+uintptr_t *return_address; // that's where the dynarec will restart when
 // going back from a C function
 
 static int32_t *SRC;      // currently recompiled instruction in the input stream
